+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+* Improve `Scroll::min_zoom`.
+    - Now minimum possible scale is `0.01.pct()`.
+    - A value under `0.01.pct()` indicates the scale needed to fit the content in the viewport.
+    - New default value is `0.fct()`.
 
 # 0.20.0